[Departures][SIRI-SM] Move some code sources
