+++ cmd/forseti/main.go
@@ -12,6 +12,7 @@ import (
 	"github.com/hove-io/forseti/internal/connectors"
 	"github.com/hove-io/forseti/internal/departures"
 	"github.com/hove-io/forseti/internal/departures/rennes"
+	"github.com/hove-io/forseti/internal/departures/sirism"
 	"github.com/hove-io/forseti/internal/departures/sytralrt"
 	"github.com/hove-io/forseti/internal/equipments"
 	"github.com/hove-io/forseti/internal/freefloatings"
@@ -19,7 +20,6 @@ import (
 	"github.com/hove-io/forseti/internal/freefloatings/fluctuo"
 	"github.com/hove-io/forseti/internal/manager"
 	"github.com/hove-io/forseti/internal/parkings"
-	"github.com/hove-io/forseti/internal/sirism"
 	"github.com/hove-io/forseti/internal/vehicleoccupancies"
 	"github.com/hove-io/forseti/internal/vehiclepositions"
 	"github.com/pkg/errors"
